[refactor] move deploy related init code to deploy package
